add the Nav to Main Page

+++ app/src/main/java/com/denny/android/idea_note/dao/NoteDao.java
@@ -6,7 +6,7 @@ import com.denny.android.idea_note.domain.NotePreview;
  * Created by hasee on 2015/6/14.
  */
 public interface NoteDao {
-    NotePreview findNoteById(int id);
+    NotePreview findNoteById(long id);
     boolean save(NotePreview note);
     boolean update(NotePreview note);
 }
